# docs/swagger.yaml
basePath: /
definitions:
...
  commoncodes.BatchCreateRequest:
    properties:
      codes:
        items:
          $ref: '#/definitions/commoncodes.CreateCommonCodeRequest'
        type: array
    type: object
  commoncodes.BatchDeleteRequest:
    properties:
      ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
  commoncodes.BatchOperationResponse:
    properties:
      failed_count:
        example: 0
        type: integer
      failed_ids:
        items:
          type: integer
        type: array
      success_count:
        example: 5
        type: integer
    type: object
  commoncodes.BatchUpdateRequest:
    properties:
      updates:
        items:
          properties:
            category:
              example: rank
              type: string
            code:
              example: MANAGER
              type: string
            description:
              type: object
            extra_payload:
              type: object
            id:
              example: 1
              type: integer
            name:
              type: object
            sort_order:
              example: 10
              type: integer
          type: object
        type: array
    type: object
  commoncodes.CategoryResponse:
    properties:
      category:
        example: rank
        type: string
    type: object
  commoncodes.CommonCodeListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/commoncodes.CommonCodeResponse'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  commoncodes.CommonCodeResponse:
    properties:
      category:
        example: rank
        type: string
      code:
        example: MANAGER
        type: string
      description:
        type: object
      extra_payload:
        type: object
      id:
        example: 1
        type: integer
      name:
        type: object
      sort_order:
        example: 10
        type: integer
    type: object
  commoncodes.CreateCommonCodeRequest:
    properties:
      category:
        example: rank
        type: string
      code:
        example: MANAGER
        type: string
      description:
        type: object
      extra_payload:
        type: object
      name:
        type: object
      sort_order:
        example: 10
        type: integer
    type: object
  commoncodes.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid category or code
        type: string
    type: object
  commoncodes.ReorderRequest:
    properties:
      orders:
        items:
          properties:
            id:
              example: 1
              type: integer
            sort_order:
              example: 10
              type: integer
          type: object
        type: array
    type: object
  commoncodes.SearchCommonCodeRequest:
    properties:
      category:
        example: rank
        type: string
      code:
        example: MANAGER
        type: string
      description:
        example: description text
        type: string
      name:
        example: Manager
        type: string
    type: object
  commoncodes.SuccessResponse:
    properties:
      message:
        example: Common code deleted successfully
        type: string
    type: object
  commoncodes.UpdateCommonCodeRequest:
    properties:
      category:
        example: rank
        type: string
      code:
        example: MANAGER
        type: string
      description:
        type: object
      extra_payload:
        type: object
      name:
        type: object
      sort_order:
        example: 10
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@knowledgecenter.io
    name: API Support
  description: Knowledge Center REST API Server
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Knowledge Center API
  version: "1.0"
paths:
...
  /common-codes:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of common codes ordered by category
        and sort_order
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.CommonCodeListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List common codes
      tags:
      - common-codes
    post:
      consumes:
      - application/json
      description: Creates a new common code. Category, code, and name are required.
        Category + code must be unique.
      parameters:
      - description: Common code data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.CreateCommonCodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commoncodes.CommonCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "409":
          description: Category and code combination already exists
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new common code
      tags:
      - common-codes
  /common-codes/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a common code by ID
      parameters:
      - description: Common Code ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete common code
      tags:
      - common-codes
    get:
      consumes:
      - application/json
      description: Retrieves a common code by its ID
      parameters:
      - description: Common Code ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.CommonCodeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get common code by ID
      tags:
      - common-codes
    put:
      consumes:
      - application/json
      description: Updates an existing common code. Category + code must be unique.
      parameters:
      - description: Common Code ID
        in: path
        name: id
        required: true
        type: integer
      - description: Common code data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.UpdateCommonCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.CommonCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "409":
          description: Category and code combination already exists
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update common code
      tags:
      - common-codes
  /common-codes/batch:
    delete:
      consumes:
      - application/json
      description: Deletes multiple common codes in a single transaction
      parameters:
      - description: Batch delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.BatchDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch delete common codes
      tags:
      - common-codes
    post:
      consumes:
      - application/json
      description: Creates multiple common codes in a single transaction
      parameters:
      - description: Batch create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.BatchCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commoncodes.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch create common codes
      tags:
      - common-codes
    put:
      consumes:
      - application/json
      description: Updates multiple common codes in a single transaction
      parameters:
      - description: Batch update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.BatchUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch update common codes
      tags:
      - common-codes
  /common-codes/categories:
    get:
      consumes:
      - application/json
      description: Retrieves all unique categories from common codes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/commoncodes.CategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all categories
      tags:
      - common-codes
  /common-codes/categories/{category}:
    get:
      consumes:
      - application/json
      description: Retrieves all common codes in a specific category, ordered by sort_order
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/commoncodes.CommonCodeResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get codes by category
      tags:
      - common-codes
  /common-codes/categories/{category}/reorder:
    put:
      consumes:
      - application/json
      description: Updates the sort_order of multiple codes in a category
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        type: string
      - description: Reorder request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.ReorderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reorder codes in a category
      tags:
      - common-codes
  /common-codes/search:
    post:
      consumes:
      - application/json
      description: Searches for common codes based on category, code, name, or description
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commoncodes.SearchCommonCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commoncodes.CommonCodeListResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commoncodes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search common codes
      tags:
      - common-codes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"